feat(type-system): add typescript
