+++ .babelrc
@@ -1,4 +1,4 @@
 {
-  "presets": ["@babel/preset-env", "@babel/react"],
+  "presets": ["@babel/preset-env", "@babel/react", "@babel/typescript"],
   "plugins": ["@babel/plugin-proposal-class-properties"]
 }
