+++ README.md
@@ -2,13 +2,13 @@
 
 [![Coverage Status](https://coveralls.io/repos/github/Luchanso/react-has-mounted/badge.svg?branch=master)](https://coveralls.io/github/Luchanso/react-has-mounted?branch=master)
 
-High Ordered Component which detect componentDidMount() call.
+High Order Component which detect componentDidMount() call.
 
 Examples:
 
 ```jsx
 import React from 'react';
-import { hasMounted } from 'react-has-mounted';
+import { withCheckMount } from 'react-has-mounted';
 
 const UserLocationInfo = ({ hasMounted }) => (
     <p>
@@ -17,7 +17,7 @@ const UserLocationInfo = ({ hasMounted }) => (
     </p>
 )
 
-export default hasMounted()(UserLocationInfo);
+export default withCheckMount()(UserLocationInfo);
 
 ```
 
@@ -27,12 +27,12 @@ React-redux and other hocs:
 import react from 'react';
 import { connect } from 'react-redux';
 import performance from 'arui-feather/performance';
-import { hasMounted } from 'react-has-mounted';
+import { withCheckMount } from 'react-has-mounted';
 import UserName from './UserName';
 
 const mapStateToProps = state => ({ name: state.user.name });
 
-export default performance(hasMounted(connect(mapStateToProps)))(UserName);
+export default performance(withCheckMount(connect(mapStateToProps)))(UserName);
 ```
 
 more clear with [compose](https://medium.com/dailyjs/react-composing-higher-order-components-hocs-3a5288e78f55):
@@ -42,14 +42,14 @@ import react from 'react';
 import { compose } from 'redux';
 import { connect } from 'react-redux';
 import performance from 'arui-feather/performance';
-import { hasMounted } from 'react-has-mounted';
+import { withCheckMount } from 'react-has-mounted';
 import UserName from './UserName';
 
 const mapStateToProps = state => ({ name: state.user.name });
 
 export default compose(
     performance,
-    hasMounted,
+    withCheckMount,
     connect(mapStateToProps)
 )(UserName);
 ```
