+++ package.json
@@ -19,6 +19,10 @@
     "@babel/plugin-proposal-class-properties": "7.2.3",
     "@babel/preset-env": "7.2.3",
     "@babel/preset-react": "7.0.0",
+    "@types/enzyme": "^3.1.15",
+    "@types/jest": "^23.3.10",
+    "@types/react": "^16.7.18",
+    "@types/react-dom": "^16.0.11",
     "babel-core": "7.0.0-bridge.0",
     "babel-jest": "23.6.0",
     "coveralls": "3.0.2",
@@ -26,7 +30,8 @@
     "enzyme-adapter-react-16": "1.7.1",
     "jest": "23.6.0",
     "react": "16.7.0",
-    "react-dom": "16.7.0"
+    "react-dom": "16.7.0",
+    "typescript": "^3.2.2"
   },
   "targets": "> 0.25%, not dead",
   "jest": {
@@ -38,5 +43,8 @@
   "peerDependencies": {
     "react": "^16.7.0",
     "react-dom": "^16.7.0"
+  },
+  "dependencies": {
+    "utility-types": "^3.2.0"
   }
 }
