+++ src/types.tsx
@@ -1,0 +1,3 @@
+export type Omit<T, K> = Pick<T, Exclude<keyof T, K>>;
+
+export type Subtract<T, K> = Omit<T, keyof K>;
